+++ docs/CHANGELOG.md
@@ -3,12 +3,13 @@
 ## 2.2.0 // TBD
 ### New features
 * Support for German language (de-de)
+* Fix spacing between multi-word abbreviations (e.g., hl. m., v. u. Z.) with respect to locale-specific spacing
 
 ### Improvements
 * Improvements in identification of wrong use of ellipsis and aposiopesis
 * Improvements in identification of wrong use of single quotes
 * Improvements in identification of name initials, e.g. J. Hus, F.X. Šalda or Ch.G.D. Lambert. Also, previous portion of the script was catching many false positives, which was reduced with this update.
-* Make en dash the default dash for Czech language 
+* Make en dash the default dash for Czech language
 
 
 ## 2.1.9 // 2018-11-11